+++ front/package.json
@@ -11,6 +11,8 @@
   "dependencies": {
     "@emotion/cache": "^11.13.1",
     "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
+    "@mui/material": "^6.1.1",
     "@mui/material-nextjs": "^6.1.1",
     "next": "14.2.13",
     "react": "^18",
